Move (most) public parts of format_arg.h to format.h
