+++ include/nanofmt/format_arg.h
@@ -16,10 +16,6 @@ namespace NANOFMT_NS::detail {
     // avoid explicitly pulling in <utility>
     template <typename T>
     T&& declval() noexcept;
-
-    // helper for c++17 and older
-    template <class T>
-    using remove_cvref_t = std::remove_cv_t<std::remove_reference_t<T>>;
 } // namespace NANOFMT_NS::detail
 
 namespace NANOFMT_NS {
@@ -86,31 +82,31 @@ namespace NANOFMT_NS {
 
 namespace NANOFMT_NS::detail {
     template <typename T>
-    struct value_type_map {
+    struct value_type_map_impl {
         using type = T;
     };
     template <>
-    struct value_type_map<signed char> {
+    struct value_type_map_impl<signed char> {
         using type = signed int;
     };
     template <>
-    struct value_type_map<unsigned char> {
+    struct value_type_map_impl<unsigned char> {
         using type = signed int;
     };
     template <>
-    struct value_type_map<signed short> {
+    struct value_type_map_impl<signed short> {
         using type = signed int;
     };
     template <>
-    struct value_type_map<unsigned short> {
+    struct value_type_map_impl<unsigned short> {
         using type = signed int;
     };
     template <>
-    struct value_type_map<decltype(nullptr)> {
+    struct value_type_map_impl<decltype(nullptr)> {
         using type = void const*;
     };
     template <typename T>
-    using value_type_map_t = typename value_type_map<T>::type;
+    using value_type_map_t = typename value_type_map_impl<std::decay_t<T>>::type;
 
     template <typename T, typename = void>
     struct has_formatter {
@@ -136,53 +132,17 @@ namespace NANOFMT_NS::detail {
 
     template <typename ValueT>
     constexpr format_value make_format_value(ValueT const& value) noexcept {
-        using T = std::decay_t<detail::remove_cvref_t<ValueT>>;
-        if constexpr (std::is_constructible_v<format_value, detail::value_type_map_t<T>>) {
-            return detail::value_type_map_t<T>(value);
+        if constexpr (std::is_constructible_v<format_value, detail::value_type_map_t<ValueT>>) {
+            return detail::value_type_map_t<ValueT>(value);
         }
-        else if constexpr (detail::has_formatter<T>::value) {
+        else if constexpr (detail::has_formatter<ValueT>::value) {
             format_value::custom custom;
-            custom.thunk = &detail::formatter_thunk<T>;
+            custom.thunk = &detail::formatter_thunk<ValueT>;
             custom.value = detail::addressof(value);
             return custom;
         }
-        else if constexpr (std::is_enum_v<T>) {
-            return static_cast<detail::value_type_map_t<std::underlying_type_t<T>>>(value);
+        else if constexpr (std::is_enum_v<ValueT>) {
+            return static_cast<detail::value_type_map_t<std::underlying_type_t<ValueT>>>(value);
         }
     };
 } // namespace NANOFMT_NS::detail
-
-namespace NANOFMT_NS {
-    /// Holds a list of N format_value objects.
-    ///
-    /// This is primarily meant to be an intermediate that holds onto values
-    /// as a temporary object, and will usually be converted to format_args.
-    ///
-    template <size_t N>
-    struct value_store {
-        static constexpr size_t size = N;
-        format_value values[N + 1 /* avoid size 0 */];
-    };
-
-    /// List of format args.
-    ///
-    /// Only use this type as a temporary value!
-    ///
-    struct format_args {
-        template <size_t N>
-        constexpr /*implicit*/ format_args(value_store<N>&& values) noexcept : values(values.values)
-                                                                             , count(N) {}
-
-        void format(unsigned index, char const** in, char const* end, buffer& buf) const;
-
-        format_value const* values = nullptr;
-        size_t count = 0;
-    };
-
-    /// Constructs a format_args from a list of values.
-    ///
-    template <typename... Args>
-    constexpr auto make_format_args(Args const&... args) noexcept {
-        return value_store<sizeof...(Args)>{detail::make_format_value(args)...};
-    }
-} // namespace NANOFMT_NS
